# backend/docs/swagger.yaml
definitions:
...
  github_com_raids-lab_crater_dao_model.GpuAnalysis:
    properties:
      command:
        description: 采集到的原始数据
        type: string
      createdAt:
        description: 自动追踪的时间戳
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      historicalMetrics:
        type: string
      id:
        type: integer
      jobID:
        type: integer
      jobName:
        type: string
      llmversion:
        type: string
      namespace:
        type: string
      phase1LLMReason:
        type: string
      phase1Score:
        description: LLM 分析结果
        type: integer
      phase2LLMReason:
        type: string
      phase2Score:
        type: integer
      podName:
        description: 原始 Kubernetes 信息
        type: string
      reviewStatus:
        allOf:
        - $ref: '#/definitions/github_com_raids-lab_crater_dao_model.ReviewStatus'
        description: 管理状态
      userID:
        type: integer
      userName:
        type: string
    type: object
...
  github_com_raids-lab_crater_dao_model.ReviewStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    format: int32
    type: integer
    x-enum-comments:
      _: 零值，被忽略
      ReviewStatusConfirmed: 已确认
      ReviewStatusIgnored: 已忽略
      ReviewStatusPending: 待审核
    x-enum-descriptions:
    - 零值，被忽略
    - 待审核
    - 已确认
    - 已忽略
    x-enum-varnames:
    - _
    - ReviewStatusPending
    - ReviewStatusConfirmed
    - ReviewStatusIgnored
...
  github_com_raids-lab_crater_dao_model.UserInfo:
    properties:
      nickname:
        type: string
      username:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Page-internal_handler_vcjob_JobResp:
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler_vcjob.JobResp'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  github_com_raids-lab_crater_internal_resputil.Response-any:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis:
    properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.GpuAnalysis'
      msg:
        type: string
    type: object
...
  ? github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_vcjob_JobResp
  : properties:
      code:
        description: 依然保持 int (ErrorCode) 类型
        type: integer
      data:
        $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Page-internal_handler_vcjob_JobResp'
      msg:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internal_handler.TriggerJobAnalysisReq:
    properties:
      jobName:
        type: string
    required:
    - jobName
    type: object
...
  internal_handler_vcjob.JobResp:
    properties:
      billedPointsTotal:
        type: number
      completedAt:
        type: string
      createdAt:
        type: string
      jobName:
        type: string
      jobType:
        type: string
      locked:
        type: boolean
      lockedTimestamp:
        type: string
      name:
        type: string
      nodes:
        items:
          type: string
        type: array
      owner:
        type: string
      permanentLocked:
        type: boolean
      queue:
        type: string
      resources:
        $ref: '#/definitions/v1.ResourceList'
      startedAt:
        type: string
      status:
        type: string
      userInfo:
        $ref: '#/definitions/github_com_raids-lab_crater_dao_model.UserInfo'
      waitingToleranceSeconds:
        type: integer
    type: object
...
  resource.Quantity:
    properties:
      Format:
        enum:
        - DecimalExponent
        - BinarySI
        - DecimalSI
        type: string
        x-enum-comments:
          BinarySI: e.g., 12Mi (12 * 2^20)
          DecimalExponent: e.g., 12e6
          DecimalSI: e.g., 12M  (12 * 10^6)
        x-enum-descriptions:
        - e.g., 12e6
        - e.g., 12Mi (12 * 2^20)
        - e.g., 12M  (12 * 10^6)
        x-enum-varnames:
        - DecimalExponent
        - BinarySI
        - DecimalSI
    type: object
...
  v1.ResourceList:
    additionalProperties:
      $ref: '#/definitions/resource.Quantity'
    type: object
...
info:
  contact: {}
  description: This is the API server for Crater, a Multi-tenant AI Model Training
    Platform based on Kubernetes.
  title: Crater API
  version: 1.0.0
paths:
...
  /v1/admin/gpu-analysis/trigger/job:
    post:
      consumes:
      - application/json
      description: 提供 Job 的 name 来启动一次完整的分析。如果该 Job 已有分析记录，则更新；否则创建新记录。
      parameters:
      - description: Job 信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/internal_handler.TriggerJobAnalysisReq'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功，返回创建或更新后的分析记录
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_dao_model_GpuAnalysis'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: 分析过程中发生错误
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: 手动触发对单个 Job 的 GPU 分析
      tags:
      - GpuAnalysis
...
  /v1/admin/vcjobs/{name}:
    delete:
      consumes:
      - application/json
      description: 管理员删除用户作业
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Admin delete the job
      tags:
      - VolcanoJob
...
  /v1/vcjobs:
    get:
      consumes:
      - application/json
      description: Get the jobs of the user by client-go
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size, 1-200
        in: query
        name: page_size
        type: integer
      - description: Sort fields
        in: query
        name: sort
        type: string
      - description: Search jobs
        in: query
        name: search
        type: string
      - description: Number of days to look back, -1 for all
        in: query
        name: days
        type: integer
      - collectionFormat: multi
        description: Job types
        in: query
        items:
          type: string
        name: job_type
        type: array
      - collectionFormat: multi
        description: Job statuses
        in: query
        items:
          type: string
        name: status
        type: array
      - description: Node name
        in: query
        name: node
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Volcano Job List
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-github_com_raids-lab_crater_internal_resputil_Page-internal_handler_vcjob_JobResp'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Get the jobs of the user
      tags:
      - VolcanoJob
  /v1/vcjobs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete the job by client-go
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Delete the job
      tags:
      - VolcanoJob
...
  /v1/vcjobs/training:
    post:
      consumes:
      - application/json
      description: Create a training job
      parameters:
      - description: CreateTrainingReq
        in: body
        name: CreateTrainingReq
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "400":
          description: Request parameter error
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
        "500":
          description: Other errors
          schema:
            $ref: '#/definitions/github_com_raids-lab_crater_internal_resputil.Response-any'
      security:
      - Bearer: []
      summary: Create a training job
      tags:
      - VolcanoJob
...
securityDefinitions:
  Bearer:
    description: 访问 /login 并获取 TOKEN 后，填入 'Bearer ${TOKEN}' 以访问受保护的接口
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"